[Improved] Use the __init__ function to initialize member variable
Let the super-class initialize the window object and then we just need to
initialize the member variable self.view.
The graphvizer_panel is now created in the __init__ function instead of in
the run function. In this way, the graphvizer_panel can be created once the
plugin has been loaded instead of waiting the first run.

Signed-off-by: Hao Lee <haolee.swjtu@gmail.com>